Fix naming of perspectives without train set
The name of a perspective was based on the directory name of the train
set. However, this causes a crash if there is no train set, such as in
case of wanting to calculate per document opinion word perplexity for a
set of documents. Therefore, the naming procedure of perspectives was
changed. It now uses the directory of the test set if there is no
training set. And if there is no training set, it uses the word UNKNOWN.