Fix bug in opinion word perplexity per document
The calculation of the opinion word perplexity per document did not use
a fixed phi_opinion, but calculated a new one. This resulted in the per
document opinion word perplexity to be the same for every perspective
(because every perspective contains the same documents). This is fixed,
by loading the appropriate phi_topic and phi_opinion from disk.